parent e723365358cd27da203cb24ede93f46c307be5b7 author Mark Walker <markw@broadinstitute.org> 1687058996 -0400 committer Mark Walker <markw@broadinstitute.org> 1689622311 -0400
Move HIGH_SR_BACKGROUND, BOTHSIDES_SUPPORT, and PESR_GT_OVERDISPERSION filters to info fields

Remove unneeded ResetFilter tasks from single sample pipeline
