+++ src/svtk/svtk/standardize/__init__.py
@@ -1,4 +1,4 @@
-from .standardize import VCFStandardizer, parse_bnd_pos, parse_bnd_strands
+from .standardize import VCFStandardizer
 from .std_delly import DellyStandardizer
 from .std_wham import WhamStandardizer
 from .std_lumpy import LumpyStandardizer
