+++ wdl/CleanVcfChromosome.wdl
@@ -741,11 +741,7 @@ task FinalCleanup {
       --chrom ~{contig} \
       --prefix ~{prefix} \
       ~{vcf} stdout \
-      | fgrep -v "##INFO=<ID=HIGH_SR_BACKGROUND" \
-      | /opt/sv-pipeline/04_variant_resolution/scripts/sanitize_filter_field.py stdin stdout \
-      | fgrep -v "##INFO=<ID=MEMBERS,Number=.,Type=String," \
-      | bgzip -c \
-      > ~{prefix}.vcf.gz
+      | bcftools annotate --no-version -x INFO/MEMBERS -Oz -o ~{prefix}.vcf.gz
     tabix ~{prefix}.vcf.gz
   >>>
